# api/openapi.yaml
openapi: 3.1.0
info:
  title: Starfly Fabrics API
  description: |
    Kubernetes-native NHI identity broker and shared signals aggregator.

    Starfly validates source credentials, evaluates OPA policy, and mints
    WIMSE-compliant JWTs via RFC 8693 token exchange.

    When mTLS is enabled, protected endpoints run on port 8694.
    Public endpoints (health, JWKS, metrics) always run on port 8693.
  version: 1.0.0
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8693
    description: Plaintext (dev mode)
  - url: https://starfly.example.com:8694
    description: mTLS (production)
paths:
...
  /v1/identity/agent:
    post:
      operationId: issueAgentIdentity
      summary: Issue agent identity
      description: |
        Issues a verifiable identity token for an AI agent (MCP, A2A, WatsonX, custom).
        Evaluates agent identity OPA policy for capability and blast radius constraints.

        Protected by mTLS when TLS is enabled.
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentIdentityRequest"
            example:
              agent_name: "code-assistant"
              platform: "mcp"
              capabilities: ["query-read", "tool-execute"]
              max_blast_radius: "workspace:dev"
              delegation_depth: 2
              metadata:
                mcp_server: "cursor-server-v1"
      responses:
        "200":
          description: Agent identity issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentIdentity"
        "400":
          description: Invalid request (missing name, invalid platform, empty capabilities)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Agent identity provider not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/signals/stream:
    post:
      operationId: createStream
      summary: Create SSF stream
      description: |
        Creates a new SSF event stream. The endpoint URL must use HTTPS
        (localhost/127.0.0.1 exempted for dev/test).

        Protected by mTLS when TLS is enabled.
      tags: [Signals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamConfig"
            example:
              iss: "starfly"
              aud: "downstream-consumer"
              events_requested:
                - "https://schemas.openid.net/secevent/caep/event-type/session-revoked"
                - "https://schemas.openid.net/secevent/caep/event-type/credential-change"
              delivery_method: "push"
              endpoint_url: "https://consumer.example.com/events"
      responses:
        "201":
          description: Stream created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "501":
          description: Signal transmitter not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStream
      summary: Delete SSF stream
      description: Removes an active SSF event stream.
      tags: [Signals]
      parameters:
        - name: stream_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Stream deleted
        "400":
          description: Missing stream_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stream not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/signals/status:
    get:
      operationId: getStreamStatus
      summary: Get SSF stream status
      description: Returns the health status of an SSF stream.
      tags: [Signals]
      parameters:
        - name: stream_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stream status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamStatus"
        "400":
          description: Missing stream_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Stream not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    StreamConfig:
      type: object
      required: [iss, aud, events_requested, delivery_method]
      properties:
        iss:
          type: string
        aud:
          type: string
        events_requested:
          type: array
          items:
            type: string
        events_delivered:
          type: array
          items:
            type: string
        delivery_method:
          type: string
          enum: ["push", "poll"]
        endpoint_url:
          type: string
          description: HTTPS endpoint for push delivery (must be HTTPS, except localhost)
    Stream:
      type: object
      properties:
        stream_id:
          type: string
        iss:
          type: string
        aud:
          type: string
        events_supported:
          type: array
          items:
            type: string
        status:
          type: string
          enum: ["enabled", "paused", "disabled"]
    StreamStatus:
      type: object
      properties:
        stream_id:
          type: string
        status:
          type: string
        subject:
          type: string
...
    AgentIdentityRequest:
      type: object
      required: [agent_name, platform, capabilities]
      properties:
        agent_name:
          type: string
          description: Human-readable agent name
        platform:
          type: string
          description: Agent platform
          enum: ["mcp", "a2a", "watsonx", "custom"]
        capabilities:
          type: array
          items:
            type: string
          description: Requested capabilities
        on_behalf_of:
          type: string
          description: Parent agent ID for delegation
        max_blast_radius:
          type: string
          description: Maximum blast radius (e.g., "workspace:dev", "namespace", "cluster")
        delegation_depth:
          type: integer
          description: Max delegation hops (0 = terminal agent)
        metadata:
          type: object
          additionalProperties:
            type: string
    AgentIdentity:
      type: object
      properties:
        agent_id:
          type: string
          description: WIMSE-format agent identifier
        token:
          type: string
          description: Signed JWT agent identity token
        platform:
          type: string
        capabilities:
          type: array
          items:
            type: string
        blast_radius:
          type: string
        expires_at:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error code (RFC 8693 compatible)
          enum:
            - "invalid_request"
            - "unsupported_grant_type"
            - "access_denied"
            - "server_error"
            - "not_implemented"
            - "not_found"
            - "method_not_allowed"
        error_description:
          type: string
          description: Human-readable error detail
...